[ADD] view for the message body
