+++ emails_align/models/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from . import mail_thread_align
+from . import mail_thread
