--- emails_align/security/ir.model.access.csv
@@ -1,7 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_mail_thread_align,access.mail.thread.align,model_mail_thread_align,base.group_user,1,1,1,1
-
-
-
-
-
